.history/requirements_20240608201930.txt
Flask==2.0.3
pybit==2.2.1
python-dotenv==0.19.2
pandas==1.3.5
ta==0.7.0